# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteCreate": {
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/AddressCreate"
          },
          "expected_time": {
            "format": "time",
            "title": "Expected Time",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/AddressCreate"
          },
          "recurrence": {
            "title": "Recurrence",
            "type": "string"
          },
          "route_type": {
            "pattern": "^(outbound|inbound)$",
            "title": "Route Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "route_type",
          "origin",
          "destination",
          "expected_time",
          "recurrence"
        ],
        "title": "RouteCreate",
        "type": "object"
      },
...
      "RoutePassangerResponse": {
        "properties": {
          "dependent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependent Id"
          },
          "dependent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependent Name"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "joined_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined At"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "pickup_address_id": {
            "format": "uuid",
            "title": "Pickup Address Id",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "route_id": {
            "format": "uuid",
            "title": "Route Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          },
          "user_phone": {
            "title": "User Phone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "route_id",
          "status",
          "requested_at",
          "user_id",
          "user_name",
          "user_phone",
          "pickup_address_id"
        ],
        "title": "RoutePassangerResponse",
        "type": "object"
      },
...
      "RouteResponse": {
        "properties": {
          "accepted_count": {
            "default": 0,
            "description": "Quantidade de passageiros aceitos na rota",
            "title": "Accepted Count",
            "type": "integer"
          },
          "active_trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ID da trip em andamento para esta rota, null se não houver",
            "title": "Active Trip Id"
          },
          "destination_address": {
            "$ref": "#/components/schemas/AddressResponse"
          },
          "estimated_duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Duração estimada planejada origem→stops→destino em minutos",
            "title": "Estimated Duration Min"
          },
          "expected_time": {
            "format": "time",
            "title": "Expected Time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invite_code": {
            "title": "Invite Code",
            "type": "string"
          },
          "max_passengers": {
            "title": "Max Passengers",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin_address": {
            "$ref": "#/components/schemas/AddressResponse"
          },
          "recurrence": {
            "title": "Recurrence",
            "type": "string"
          },
          "route_type": {
            "title": "Route Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/StopResponse"
            },
            "title": "Stops",
            "type": "array"
          },
          "total_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Distância total planejada origem→stops→destino em km",
            "title": "Total Distance Km"
          }
        },
        "required": [
          "id",
          "name",
          "route_type",
          "status",
          "recurrence",
          "expected_time",
          "invite_code",
          "max_passengers",
          "origin_address",
          "destination_address"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API para gestão de transporte escolar (VanGo)",
    "title": "VanGO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/routes/": {
      "get": {
        "description": "Retorna todas as rotas cadastradas pelo motorista, independente de status.",
        "operationId": "list_routes_routes__get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteResponse"
                  },
                  "title": "Response List Routes Routes  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar todas as rotas do motorista",
        "tags": [
          "Routes"
        ]
      },
      "post": {
        "description": "Cria uma rota para o motorista com origem, destino, horário e recorrência. Gera automaticamente o código de convite e salva com status inativa.",
        "operationId": "create_route_routes__post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Criar uma nova rota",
        "tags": [
          "Routes"
        ]
      }
    },
...
    "/routes/me": {
      "get": {
        "description": "Retorna todos os vínculos ativos (pending + accepted) do usuário autenticado, incluindo vínculos em que ele é guardian de um dependente. Ordenado por joined_at desc. Usado pela home do passageiro.",
        "operationId": "list_my_routes_routes_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PassangerRouteResponse"
                  },
                  "title": "Response List My Routes Routes Me Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar minhas rotas (home do passageiro)",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
...
    "/routes/{route_id}/me": {
      "get": {
        "description": "Retorna o detalhe completo da rota do ponto de vista do passageiro autenticado. Aceita dependent_id opcional quando guardian está vendo em nome do dependente. Não expõe invite_code/max_passengers/driver_id. Erros: 404 se a rota não existir, 403 se o usuário não tiver vínculo ativo com a rota.",
        "operationId": "get_my_route_detail_routes__route_id__me_get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dependent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dependent Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassangerRouteDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalhe da rota para o passageiro",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
    "/routes/{route_id}/passangers": {
      "get": {
        "description": "Lista vínculos de passageiros da rota. Aceita filtro opcional por status (pending, accepted, rejected). Sem filtro retorna todos.",
        "operationId": "list_passangers_routes__route_id__passangers_get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutePassangerResponse"
                  },
                  "title": "Response List Passangers Routes  Route Id  Passangers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar passageiros/solicitações da rota",
        "tags": [
          "RoutePassangers"
        ]
      },
      "post": {
        "description": "Passageiro (ou guardian em nome de dependente) solicita entrada na rota informando schedules (dias + endereço de embarque). Cria vínculo pending.",
        "operationId": "join_route_routes__route_id__passangers_post",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePassangerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Solicitar entrada em rota",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
    "/routes/{route_id}/passangers/me": {
      "delete": {
        "description": "Passageiro sai da rota. Aceita dependent_id opcional quando guardian está saindo em nome do dependente. Bloqueado se rota está em andamento.",
        "operationId": "leave_route_routes__route_id__passangers_me_delete",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dependent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dependent Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sair de uma rota",
        "tags": [
          "RoutePassangers"
        ]
      },
      "patch": {
        "description": "Substitui completamente os schedules atuais do passageiro pelos fornecidos. Aceita dependent_id opcional para guardian em nome de dep.",
        "operationId": "update_schedules_routes__route_id__passangers_me_patch",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dependent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dependent Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchedulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePassangerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar schedules de participação",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
    "/routes/{route_id}/passangers/{rp_id}": {
      "delete": {
        "description": "Remove um passageiro da rota. Bloqueado se a rota estiver em andamento.",
        "operationId": "remove_passanger_routes__route_id__passangers__rp_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rp Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remover passageiro da rota",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
    "/routes/{route_id}/passangers/{rp_id}/accept": {
      "post": {
        "description": "Aprova uma solicitação pending. Valida ownership, bloqueio de rota em andamento e capacidade máxima do veículo.",
        "operationId": "accept_request_routes__route_id__passangers__rp_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rp Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePassangerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aceitar solicitação de passageiro",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
    "/routes/{route_id}/passangers/{rp_id}/reject": {
      "post": {
        "description": "Recusa uma solicitação pending. Valida ownership e bloqueio de rota em andamento. Não valida capacidade.",
        "operationId": "reject_request_routes__route_id__passangers__rp_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rp Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePassangerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recusar solicitação de passageiro",
        "tags": [
          "RoutePassangers"
        ]
      }
    },
...
  }
}